requirements.txt
streamlit
pypdf2
langchain
python-dotenv 
faiss-cpu
openai
huggingface_hub
langchain==0.0.123
langchain-community
InstructorEmbedding
sentence_transformers
transformers